+++ packages/telestion-client-common/src/components/pages/login-page/index.ts
@@ -1,1 +1,6 @@
+export * from './login-form';
+
+export * from './login-description';
+export * from './login-logo';
 export * from './login-page';
+export * from './login-title';
