+++ packages/telestion-client-common/src/components/pages/login-page/login-logo.tsx
@@ -1,0 +1,16 @@
+import { Image } from '@adobe/react-spectrum';
+import { useLogo } from '../../contexts/app-logo-context';
+
+export function LoginLogo() {
+	const appLogo = useLogo();
+
+	return (
+		<Image
+			src={appLogo}
+			alignSelf="center"
+			alt="Application logo"
+			width="size-1200"
+			height="size-1200"
+		/>
+	);
+}
