+++ packages/telestion-client-common/src/components/pages/login-page/login-title.tsx
@@ -1,0 +1,7 @@
+import { Heading } from '@adobe/react-spectrum';
+import { useTitle } from '@wuespace/telestion-client-core';
+
+export function LoginTitle() {
+	const title = useTitle();
+	return <Heading level={2}>{title}</Heading>;
+}
